growup_marathon_beta_v2_1_6_nocustom/requirements.txt
aiogram==3.25.0
python-dotenv==1.0.1
aiosqlite==0.20.0
apscheduler==3.10.4
pytz==2025.2
